requirements.txt
telethon
python-dateutil==2.8.2
python-dotenv==1.0.0